+++ CHANGELOG.md
@@ -33,6 +33,8 @@ All Notable changes to `League\Uri` will be documented in this file
 - `Uri::withUsername` returns a new `Uri` instance with the updated username component.
 - `Uri::withPassword` returns a new `Uri` instance with the updated password component.
 - `UriTemplate` implements the `Stringable` interface
+- `UriTemplate::expandToUri` and `UriTemplate::expandToUrl`
+- `UriTemplate::expandToUriOrFail` and `UriTemplate::expandToUrlOrFail`
 - Dependency to `psr/http-factory` package which is required for the package.
 - Support for `Uri\Rfc3986\Uri` and `Uri\Whatwg\Url`